Clean up code. IMU and ESC interfaces self contained.

+++ QuadFC/include/control_algorithm.h
@@ -9,8 +9,9 @@
 #ifndef CONTROL_ALGORITHM_H_
 #define CONTROL_ALGORITHM_H_
 
-#include "main.h"
+#include "main_control_task.h"
 #include "debug_macros.h"
+#include "my_math.h"
 
 #define CTRL_P_LIMIT 1000
 #define CTRL_I_LIMIT 9000  
@@ -19,8 +20,8 @@
 
 
 
-void calc_control_signal_angle_pid(int32_t motor_setpoint[], uint8_t nr_motors, control_values_pid_t *parameters_pid, control_values_pid_t *ctrl_error, state_data_t *state, state_data_t *setpoint, control_signal_t *ctrl_signal);
-void calc_control_signal_rate_pid(int32_t motor_setpoint[], uint8_t nr_motors, control_values_pid_t *parameters_pid, control_values_pid_t *ctrl_error, state_data_t *state, state_data_t *setpoint, control_signal_t *ctrl_signal);
+void calc_control_signal_angle_pid(int32_t motor_setpoint[], uint32_t nr_motors, control_values_pid_t *parameters_pid, control_values_pid_t *ctrl_error, state_data_t *state, state_data_t *setpoint, control_signal_t *ctrl_signal);
+void calc_control_signal_rate_pid(int32_t motor_setpoint[], uint32_t nr_motors, control_values_pid_t *parameters_pid, control_values_pid_t *ctrl_error, state_data_t *state, state_data_t *setpoint, control_signal_t *ctrl_signal);
 void control_allocation_quad_x(control_signal_t *ctrl_signal, int32_t motor_setpoint[]);
 void limit_value(int32_t *val, int32_t limit);
 #endif /* CONTROL_ALGORITHM_H_ */
