+++ QuadFC/src/control_algorithm.c
@@ -24,7 +24,7 @@
 #include "control_algorithm.h"
 
 
-void calc_control_signal_angle_pid(int32_t motor_setpoint[], uint8_t nr_motors, control_values_pid_t *parameters_pid, control_values_pid_t *ctrl_error, state_data_t *state, state_data_t *setpoint, control_signal_t *ctrl_signal)
+void calc_control_signal_angle_pid(int32_t motor_setpoint[], uint32_t nr_motors, control_values_pid_t *parameters_pid, control_values_pid_t *ctrl_error, state_data_t *state, state_data_t *setpoint, control_signal_t *ctrl_signal)
 {
     
      static control_signal_t error_d_last;
@@ -81,7 +81,7 @@ void calc_control_signal_angle_pid(int32_t motor_setpoint[], uint8_t nr_motors, 
 }
 
 
-void calc_control_signal_rate_pid(int32_t motor_setpoint[], uint8_t nr_motors, control_values_pid_t *parameters_pid, control_values_pid_t *ctrl_error, state_data_t *state, state_data_t *setpoint, control_signal_t *ctrl_signal)
+void calc_control_signal_rate_pid(int32_t motor_setpoint[], uint32_t nr_motors, control_values_pid_t *parameters_pid, control_values_pid_t *ctrl_error, state_data_t *state, state_data_t *setpoint, control_signal_t *ctrl_signal)
 {
     
     static control_signal_t error_d_last;
